chore(stories): move stories to a separated folder
