+++ __stories__/Popup.stories.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Meta, Story } from '@storybook/react';
-import Popup, { Props } from './index';
-import './Popup.css';
+import Popup, { Props } from '../src';
+import '../src/index.css';
 
 const meta: Meta = {
   title: 'Popup',
@@ -15,7 +15,7 @@ const Template: Story<Props> = (args) => (
   <Popup
     {...args}
     toggler={
-      <button type="button">Toggler</button>
+      <button type="button" style={{ marginLeft: '50%' }}>Toggler</button>
     }
   >
     <p>Lorem ipsum dolor sit, amet consectetur adipisicing elit. Est, magnam?</p>
